[Refactor] Expose ASU transport provider factory
