+++ ucm/transport/kv/asu/trans/src/aicpu_trans_provider.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "trans_provider.h"
+#include "asu_transport/trans_provider.h"
 
 namespace UC::ASU {
