+++ ucm/transport/kv/asu/trans/src/aiv_trans_provider.h
@@ -25,7 +25,7 @@
 
 #include <memory>
 #include "aiv_transport/aiv_transport.h"
-#include "trans_provider.h"
+#include "asu_transport/trans_provider.h"
 
 namespace UC::ASU {
