+++ ucm/transport/kv/asu/trans/src/asu_transport_impl.cpp
@@ -29,17 +29,9 @@
 #include <string>
 #include <thread>
 #include <utility>
-#ifdef UCM_ASU_ENABLE_AICPU_PROVIDER
-#include "aicpu_trans_provider.h"
-#endif
-#ifdef UCM_ASU_ENABLE_AIV_PROVIDER
-#include "aiv_trans_provider.h"
-#endif
 #include "asu_transport/asu_transport.h"
+#include "asu_transport/trans_provider.h"
 #include "connection_manager.h"
-#ifdef UCM_ASU_ENABLE_FAKE_PROVIDER
-#include "fake_trans_provider.h"
-#endif
 #include "logger.h"
 #include "transport_config_parser.h"
 
@@ -80,40 +72,8 @@ Status AsuTransportImpl::Init(const TransportConfig& config)
     }
 
     if (!transProvider_) {
-        switch (config_.providerType) {
-            case TransProviderType::AICPU:
-#ifdef UCM_ASU_ENABLE_AICPU_PROVIDER
-                transProvider_ = std::make_unique<AICPUTransProvider>();
-                break;
-#else
-                return Status::Error(
-                    StatusCode::UNSUPPORTED,
-                    "AICPU trans provider is not built; enable BUILD_UCM_ASU_PROVIDER_AICPU");
-#endif
-            case TransProviderType::FAKE:
-#ifdef UCM_ASU_ENABLE_FAKE_PROVIDER
-                transProvider_ =
-                    std::make_unique<FakeTransProvider>(MakeFakeTransProviderConfig(config_));
-                break;
-#else
-                return Status::Error(
-                    StatusCode::UNSUPPORTED,
-                    "FAKE trans provider is not built; enable BUILD_UCM_ASU_PROVIDER_FAKE");
-#endif
-            case TransProviderType::AIV:
-#ifdef UCM_ASU_ENABLE_AIV_PROVIDER
-                transProvider_ = std::make_unique<AIVTransProviderAdapter>(config_.deviceId);
-                break;
-#else
-                return Status::Error(
-                    StatusCode::UNSUPPORTED,
-                    "AIV trans provider is not built; enable BUILD_UCM_ASU_PROVIDER_AIV and set "
-                    "ASU_AIV_PROVIDER_ROOT");
-#endif
-            case TransProviderType::UNSUPPORTED:
-                return Status::Error(StatusCode::UNSUPPORTED,
-                                     "ASU trans provider backend is not supported");
-        }
+        auto status = CreateTransProvider(config_, transProvider_);
+        if (!status.ok()) { return status; }
     }
     if (!transProvider_) {
         UC_ERROR("AsuTransportImpl::Init: TransProvider is null");
@@ -476,7 +436,7 @@ void AsuTransportImpl::CompletionLoop()
     }
 }
 
-void AsuTransportImpl::SetTransProvider(std::unique_ptr<TransProvider> provider)
+void AsuTransportImpl::SetTransProvider(std::shared_ptr<TransProvider> provider)
 {
     transProvider_ = std::move(provider);
 }
