+++ ucm/transport/kv/asu/trans/src/fake_trans_provider.h
@@ -28,7 +28,7 @@
 #include <mutex>
 #include <unordered_map>
 #include "asu_transport/asu_transport.h"
-#include "trans_provider.h"
+#include "asu_transport/trans_provider.h"
 
 namespace UC::ASU {
