+++ ucm/transport/kv/asu/trans/src/transport_task_executor.cpp
@@ -57,7 +57,7 @@ Status CopyDeviceToHost(const ScatterGatherEntry& sge, void* host, std::size_t s
 
 TransportTaskExecutor::TransportTaskExecutor(
     const TransportConfig& config, IoScheduler& ioScheduler,
-    const std::unique_ptr<TransProvider>& transProvider, BufferManager& sendBufferManager,
+    const std::shared_ptr<TransProvider>& transProvider, BufferManager& sendBufferManager,
     BufferManager& flagBufferManager, const std::unique_ptr<ProtocolManager>& protocolManager,
     const std::unique_ptr<ConnectionManager>& connectionManager,
     std::atomic<std::uint16_t>& nextRequestCid, std::mutex& registeredRegionsMu,
